Pin psr/http-message to 1.0.1
When using the `implements` keyword on a class in PHP, all required
methods of the interface must be implemented.
With the dawn of type hints in PHP, this feature has become even more
useful. *Especially* if you want any implementers of your interface to
VERY strictly follow the procedures.
A simple `declare(strict_types=1)` at the top of your interface's file
can potentially ruin someone's day. It means that if you have the
following method in your interface: `function a(string $b): int`, any
implementers MUST copy ALL typehints, or PHP will throw an exception.

MediaWiki fails to do this in one specific instance. Here's the links:
- https://github.com/wikimedia/mediawiki/blob/d1d58868d28f52a84cd8541ac868a4ce56917aba/includes/http/MWCallbackStream.php#L49
- https://gerrit.wikimedia.org/r/plugins/gitiles/mediawiki/core/+/refs/heads/master/includes/http/MWCallbackStream.php#49

Now, let's see where the problem is. This commit to `psr/http-message`
added type hints and enabled strict_types:
https://github.com/php-fig/http-message/commit/5cd5ad79255dfa7243fdbc704c6f6b7374f1c07a
As a result, MediaWiki 1.39.3 contains this error. There is a closed
issue about it on WikiMedia Phabricator:
https://phabricator.wikimedia.org/T335073
...which, in turn, mentions this patchset:
https://phabricator.wikimedia.org/T333993

The description on that second issue there leads me to believe the
MediaWiki team are against going with the times -- which would entail
adding *five characters* to the end of the function implementation
(namely `: int`), and are instead explicitly pinning this library to a
version that does not contain type hints for MediaWiki version 1.39.4,
which has yet to release. Very lazy solution, guys. Sadly, I have no
choice but to comply with this change until 1.39.4 releases.